feature(edit post): add edit post
[Starts #174459946]
